// katana-slice-manager/src/app/models/interfaces/vim-registration-form.interface.ts
export interface VimRegistrationFormModel {
  id: string;
  name: string;
  authUrl: string;
  username: string;
  password: string;
  adminProjectName: string;
  location: string;
  type: string;
  version: string;
  description: string;
  infrastructureMonitoring: string;
  securityGroups: string;
}